Rename BaseDAOTest to AbstractDAOTest and introduce an actual BaseDAO test case.
This was confusing; BaseDAOTest suggested it was testing the functionality of the BaseDAO class, when it was just a test helper to test DAOs. We should have a place to test the core DAO functionality too, that why the "new" BaseDAOTest is now an actual test case where unit tests for that class should go.