Invalidate redirect on client status codes
We cannot rely on certain status codes to be returned since we are migrating to different CDNs. To be safe, we should invalidate all redirects whenever we see a client error, which will put the responsibility on the head task to either provide a new URL or mark something as unavailable.