Bring back functionality that rethrows fatal errors
fixes #1823

This is admittedly hacky, but I couldn’t find another working solution. This basically tricks RxJava into thinking there’s no error handler defined for the given exception that we want to be fatal. I will keep following up with the core team for how to make this more configurable in the library.