work in progress towards extending PlaybackSessionEvent to track ads
There was handling for UnsupportedEncodingExceptions in EventloggerAnalyticsProvider that
seemed unreachable, so I removed it.
Temporarily made PlaybackSessionEvent non-final to allow mocking to explore new
interfaces on it. I think longer term solution is a proper Blueprint for it?
New code is effectively unreachable because PlaybackSessionEvent never admits
to playing an ad.

refs #2011