Replace ReplaySubject of size 1 with BehaviorSubject for replay queues
I don’t think the solution using replay subject worked correctly. I saw double emissions of events. Not sure why we didn’t use BehaviorSubject to begin with, as it’s meant to do exactly that, but I vaguely remember it had to do with requiring a default event or something.