Do not play a track if its related id list is empty
This bug happened when, for example, you try to play a playlist track for a playlist which has no track on the database. More details on the related issue.

The consequence of this fix is that when you try to play a track on an empty playlist, the click action on a track will not do anything.

Fixes #2075