Uses playlist DB tracks to playback on playlist header toggle button click
Now when the playlist header toggle button is pressed, we use the playlist tracks on the playlist instead of using the cached model on the fragment.

One improvement could be made here to refresh the playlist screen when we receive a IllegalStateException on ShowPlayerAfterPlaybackSubscriber, which happens where there is mismatch between the tracks on the screen and the ones on the database.

Fixes #2143