Integrate a version of Google Play Services with Advertisement ID.
This is a pre-requisite for Adjust SDK integration. This integrates the lowest version of GPS that contains this guy in order to have a light weight impact.

Since we can’t update the Android version (because of the apache dependency) this won’t build unless we setup a specific proguard configuration to ignore warning in the gms package.

Refs #2383