Remove the clean up logic for the DB.
The clean up logic complexify the code and if tied to the deprecated ScContentProvider. While moving forward, this logic is a pain point. Considering we do not know it is useful, the decision was made to remove it and add metrics around the DB size. Using the latter we may decide to re-introduce it in the future. As a side note, currently, iOS does not have any similar logic.  

Refs #2612