Fix NPE causing a crash when a user try so sign up and is blocked.
The Onboarding activity used to set up the movement method on the alert dialog. But since the text view id changed, provoking a `null` view to be dereferenced and as a consequence the app crashed. 

This line was removed because since the Material dialogs lib was integrated to the app, it is not mandatory anymore - it is the default setup. Also it actually did not make any sense with the displayed text in this context.