Delay the activity result actions to the next safe lifecycle method: onResumeFragments()
Fixes #3056

The exception was actually being thrown because we are trying to perform a fragment transaction `onActivityResult` - which happens before `onResume()` - so we may incur in a "activity state loss" situation [1]. if we show a fragment `onActivityResult`

Instead, the actions (which include our google fragment transactions) can be done on the next safe lifecycle method, `onResumeFragments()`. That is the fix I will post shortly in a PR.

Why `onResumeFragments()` is used instead of plain `onResume()`?
> Dispatch onResume() to fragments. Note that for better inter-operation with older versions of the platform, at the point of this call the fragments attached to the activity are not resumed. This means that in some cases the previous state may still be saved, not allowing fragment transactions that modify the state. To correctly interact with fragments in their proper state, you should instead override onResumeFragments() 
 
 
[1] http://www.androiddesignpatterns.com/2013/08/fragment-transaction-commit-state-loss.html
[2] https://developer.android.com/reference/android/support/v4/app/FragmentActivity.html#onResume()