Fix the openDrawer test
Since system menu is a modal dialogue, it covers the drawer completely so Robotium is not able to see it in the viewHierarchy. Better thing to do would be to open and close system menu and evaluate visibility then. 
I wonder if this test makes any sense since it's a android's standard component, should we even test those?