Add the network type and operator name as part of the debug message when logging failed.
This because some users can’t resolve SoundCloud hosts even though they are connected to the network. One user from the beta community mentioned solving the problem when changing is DNS. Unfortunately, logging the DNS would be too much of a hack on Android but Conrad from Traffic suggested we log the connexion type and operator name. 

Closes #3359