Implementation of tracking for play related tracks.
In order to implement this tracking, playNewQueue had to accept a PlayQueue instead of a List<Urn>. This PR first changes the PlaybackStrategy playNewQueue interface needed to accept a PlayQueue instead of a List<Urn>. We then refactored the methods calling playNewQueue to use a PlayQueue. Third, we created a way for the playTrackWithRecommendations method to build a play queue with PlayQueueItems that contain the appropriate sources & page_name.

We now have PlayQueue builders for default, shuffled, and recommmender play queues that can be used to easily make PlayQueues from a List<Urn>.

https://soundcloud.atlassian.net/browse/STAT-89