Fix the initial drawer item selection.
The fragment hook in the drawer list view to select the initial item on resume. Unfortunately, the position in the list view is not the same that the one in the adapter for a given item. This because the fragment set a header that occupies the first position of the list view but is absent from the adapter. 

Fixes #3585