Simplify player presenter binding + move tests
In preparation for further player work, I am trying to simplify the parameters that we use for binding by using viewModels in our adapter. This way we don’t have to have excess information that we do not care about in the presenters, but can instead just use a base class for the view models.

I also started the arduous task of moving the tests over.