Bugfix: When stations are requested from the service, don't wipe out local column LAST_PLAYED_TRACK_POSITION.
LAST_PLAYED_TRACK_POSITION is a column used only by the client to start stations from the correct start position in the play queue. It is not synced with the service. Previously, the when the service returned data for a station, we would reset this value in the two stations storage commands. To avoid this, the default value in the database is NULL, and ensure that the column is not overwritten in the commands meant to store stations coming from the backend.