for syncs that error, wait until next time to try again
When a stream sync failed, the code was re-trying
immediately with the idea that the future_href could
have gone stale and caused the sync failure. But
this code is causing self-DOS during outages. If
Kraken or Stitch are having trouble, they have more
trouble because Android retries any 500 right away.

Desk checked this with Tom before committing.

MC-358