Simplify new profile layout, fix action bar
We do not use any of the CollapsingToolbarLayout specific functionality (except the content scrim, which arguably is totally unnecessary since we do not use full bleed images). This removes that in favour of a pure AppBarLayout, which also allows us to use normal title functionality.

As part of this, also fixing the toolbar to not have search or overflow functionality.

Fixes #3922