Display snippetized waveform and limit progress
https://soundcloud.atlassian.net/browse/PLAYBACK-772

This issue does not have the proper comps in it. Cannot find the one that does.

Basically, we have to make the UX for snippets more friendly and it should be obvious that the user is not listening to a full track, but rather part of one. We do this by returning the snippet + full durations from the api, and using them to change the way the waveform is displayed, only showing partial progress, and allowing partial scrubbing