Make sure the track context is up-to-date
We found a bug with the stations context: When you play a track, then
start the station from that track, the context would not appear. The
problem has to do with caching, and that we don’t rebind views when we
load the views from the cache.

We now want to rebind the views every time, but this can cause
flickering. Therefore, we only clear the previous item view if it
wasn’t pulled from the cache.

Furthermore, the playback state transition caused flickering as
well. To reduce this, we’ve debounced the effect.