Load first page of Activities full stack
This is not a full implementation, but all major collaborators are in place.
This merely loads the first page from local storage. More complicated sync based logic is to come in a follow up PR.

Add ActivitiesFragment and dummy presenter

Rename existing ActivitiesAdapter to LegacyActivitiesAdapter

Add new ActivitiesAdapter

Wire up Activities presentation with ops and storage