Atomize subscriber access in ResultReceiverAdapter
Fixes #4284

This was caused by the subscriber field being nulled out from another
thread while the adapter was in between making callbacks to the same
subscriber.

By using an AtomicReference instead, we reduce the opportunity for the
reference to become null to a single location, so now it should either
perform all callbacks or none at all.