Move deep link/permalink resolution to api-mobile
Urn resolver returns Optional instead of throwing exception

Migrate ResolveOperations to api-mobile

Migrate link resolution to new /resolve endpoint

This removes handling of URNs. For unknown reasons, we treated URNs as
deeplinks.

Clean up resolve acceptance tests

Fix the case where we circle back to resolve with a URN

This is super weird, but I didn’t feel like refactoring all of this.
Basically what happens here is that we get some transparent URI that we
resolve to a URN, then we decide we must log the user in, tunnel that
URN through OnboardActivity and back to the resolver… only then we
start the Activity.

Could be simplified by logging the user in FIRST, then resolving and
going straight to the target activity?? Maybe I’m missing something,
but could be that the permalink page impression tracking is the cause
for why we need to do it this way?

Better test coverage

Remove the now unused UrnResolver class