Local playlist removal
This is a combination of 17 commits:
Presentation for deletable playlists

Exclude deleted playlists from SoundView

Implement markForRemoval for posted playlists.

Add the remove playlist operations & trigger it from the dialog.

Emit a playlist deleted event and update the collections screen.

Go back in the stack when the current playlist presented in PlaylistDetailFragment is removed.

Enabled DeleteMyPlaylist acceptance tests.

Update the content provider to prefix the field “remove at” on likes and avoid conflicts with playlists.

Update Delete Playlist acceptance test to use a dedicated user and avoid interfering when running tests in parallel.