use https in media player
We see users having playback problems when they are on mobile data, and the working theory is that they would be solved by using secure streaming endpoints. This enables that for MediaPlayer, which we still use as a fallback, while we do further work on Skippy to enable it to fallback as well.

- also removed unused variable