Rxification of play state chain
This is a small refactor in prep for blocked tracks logic.

I also fixed a bug where we would try to play the current track even if it did not advance. This would probably lead to replaying the same track again in some edge cases (like when we failed to fetch recommendations).