Publish stop reason with concurrent stream reason
I extracted the logic for determining stop reasons to a StopReason provider, which is now a singleton, and will remember if it is about to be stopped because of concurrent streaming.

I am not proud of this solution. If you think it should be done a different way, please tell me how. I don’t believe this information should be encoded in our playback apis, and this is the least ugly solution I could come up with.

https://soundcloud.atlassian.net/browse/PLAYBACK-722