Offline content refactoring, it aims to fix these 3 points.
1/ Determine the collections of tracks offline content status based on all the track status.

To determine the state of a collection, the service used to compute it based on the the current Download Queue state. This was not always accurate, for instance when a track was removed, the service could not tell if the related playlists were downloaded, requested or unavailable. 


2/ Move all offline content events to the OFFLINE_CONTENT_CHANGED. 

Using the ENTITY_STATE_CHANGED for offline content proved to introduce some complexity (since 2 queues would send this offline content event) and in the end not needed. All offline content events now goes to OFFLINE_CONTENT_CHANGED. 


3 / Load tracks playlists before refreshing the policies, as part of the loadOfflineContent operation.

This operations was done in the OfflineContentController when starting the service. It did not covered some cases such as when the service restart after the device has been offline.