- Prevent the offline sync service to loop infinitely - Optimize by a lot the delay when syncing collections of big accounts - Sync offline content automatically
The offline service used to both react and trigger sync events which could trigger infinite loops. It was specifically a problem for empty playlists since we sync those no matter what.

Syncing the playlists tracks was moved out of the offline operations down to the syncer. This fix the loop issue + optimize the delay when triggering offline collections + add the support for syncing automatically offline content. 

https://soundcloud.atlassian.net/browse/COLLECTION-63