Introduce UNDEFINED Plan value
We were previously assuming FREE if we had no knowledge of the
current/previous plan state. This is problematic since it breaks the
upgrade/downgrade logic, since transitions from an unknown state to a
known state were account as plan upgrades or downgrades.

We now explicitly exclude the UNDEFINED state from such checks, so we
do not assume anything based on unknown plan state.