Fix DailyUpdateService initial delay
We were accidentally setting the initial update delay as a relative
period rather than an absolute timestamp, which meant that we would
always set it to a value in the past, thus triggering the update
service immediately after the alarm had been set.

This meant that when first signing into the app, we would update
policies and configuration multiple times.