Only show sync notifications for user triggered evts
https://soundcloud.atlassian.net/browse/COLLECTION-205

I simplified the story a bit by dividing things that trigger downloads into 2 categories;

1) The user toggling the save offline setting on a collection
2) everything else

I decided it was expected that the user would see the notification for the result of the download right after they trigger it, so they can confirm that yes, something happened. I also think its unnecessary to do it in other cases, as its basically assumed that our app will work as intended (meaning we will eventually download it).