Use api-mobile /me endpoint during login
This might stabilise some acceptance tests, or at least make them easier to debug.

Take note, I did not go all the way with this. We still use the old endpoint in the syncer. That is up next.

Also, this throws away the configuration that we get back from the /me endpoint. In the future, we should use this to update the local config and save a request