Do not advance tracks if current track didn’t end
Previously, we would take the signal of a track ending as grounds to skip to the next track. This is fine in most cases, however if we are in the middle of manually changing tracks, this may result in skipping a track. This check should fix that (and maybe stabilise some tests).

Fixes #5007