Don't error out when awaiting pending plan with no pending plan
refs #5024

What happened was that due to the other crash we found (the broken
animation on the resubscribe screen), our offboarding Activity
terminated abnormally, but the flow completed internally, thus removing
the pending plan change. When Android then resurrected the previously
visited Activity before the crash, it was in an illegal state since the
flow had already terminated.