Fix chrome cast not advancing tracks on change
Because of a refactor, the logic that tells the chrome cast to play the next track was not working. It was checking that the new track from a track change event was different from the one in the play queue manager, which is always true. What we want is to check whether it is different from the last one we knew about. Unfortunately this adds some more state, but not much.

Fixes #5163

Also removed unnecessary subscription