Remove font padding for numbers
I do not know why the font is measured different in Android N. This fixes the problem for the timestamps though. We need to keep an eye out for more text mismeasurements, and if we find them, get to the bottom more.

I tried a lot, switching fonts, using the Calligraphy library, nothing mattered.

Fixes #5166