Do not unsubscribed from playCurrent in track change
There was a piece of logic that was added during a hot fix which effectively unsubscribes from the playCurrent function whenever there is a track change. This is wrong, as when we initially start a playQueue, we call playCurrent before getting a track change.

Removing unsubscribe fixes this problem. This is how it behaved previously. This was not seen during testing as it is essentially a race condition.