Fix bug where tracks could not be clicked in list
There was a bug where, if you were displaying a mixture of blocked and
unblocked tracks in a list, and you were using a recycler adapter,
sometimes random playable tracks could not be clicked. This was because
we were not properly "fixing up" a previously-blocked track view when it
became attached to a track that is not blocked.