[CREATORS-1758] Landscape support for user profile
This adds support for landscape mode and larger displays for the new
user profile. The top section of spotlighted tracks and playlists are
split into multiple columns, while all the other sections are
full-width.

This was accomplished using the StaggeredGridLayout, and using the
'setFullWidth' layout param available on the recycler item adapters.
Only spotlight items should not be full width.