[CREATORS-1668] Add tracking tests for new profile (#5329)
* [CREATORS-1668] Add tracking tests for new profile

This adds some new acceptance tests that target the tracking specs for
th new profile. These tests were kind of annoying to write, due to the
fact that the user profile is a flat list made only to _look_ like there
is a hierarchy of sublists. This means that the test runner has to
scroll until it finds a header element, and then try to find the next
playable thing underneath that element. This seems to be the most
reliable way to accomplish this task.

* apply feedback

* refactor UserScreens to share more code

* rename methods and local variables

* rename methods for consistency