Added background syncing support for new Syncables
+ Guillaume

Added code to support back off, exception handling, and multiple jobs per request to the sync service. This allows us to perform background syncing through the sync adapter.

This is behind a feature flag, and the synced still uses the old paths, until we have had a chance to port all of the collections to syncables. This will be a separate PR.

Also moved some stuff around.