Fix Referrer, report referrer to crashlytics (#5649)
Turns out we were removing the referrer from the intent in the foreground controller, after we had tracked the screen. I fixed this by not removing it, but instead storing the fact that we had already tracked it. It is probably not good practice to remove things from intents in general, as there is no guarantee it is not used elsewhere.

I also added this information to our crashlytics event. I do not know if I added it properly, so would appreciate some feedback on that