Add DevelopmentMenu from Feature returned from server (#5683)
* Add DevelopmentMenu from Feature returned from server

See https://soundcloud.atlassian.net/browse/PLAYBACK-1335

This PR adds a new feature name "development_menu" which returns true for cloud peeps. The draw is not displayed until we make the successful call to the server i.e. there is no development draw by default.

Application logic has been removed from `BaseLayoutHelper` and now lives in the presenter.

Scope:

MainTabsPresenter doesn't have any unit tests so the actual inflation of the drawer is not tested. I can create a test for this presenter if we think unit testing is the best way to do it.

`BaseLayoutHelper` can contain only static methods and does not need to be injected. I didn't make these changes on this PR as it requires changes to nearly every activity (20+ classes)

DNM as a change to the Mobile-API is required first

* Fix styles

* Add default of drawer visible if debug build