Do not stop playback if I caused concurrent streaming (#5746)
We still have problems with regards to stopping our own plays. Andrea and I believe this has something to do with having different flavours installed on the same phone, or potentially due to multiple registrations. This is a failsafe, that checks the token that caused the problem, and does not stop if that token is ours.