Make reposts use Toasts instead of snackbars (#5749)
The way we repost, and give feedback for reposts is super complicated. For instance, if you repost without internet, we actually end up with

- local database write
- event bus event with change
- try to push and fail
- local database revert
- event bus event with original state

This is ridiculous, and leads to things like showing the fact that you “unposted” something. That you were trying to repost.

This does not fix that, it just goes back to toasts until we fix it properly.