Fix crop storage permission (#5796)
* Ask for write permission before we use image picking

This is a bit like using a steak knife when we could probably use a Scalpel, but given the state of our release, its better to be safe than sorry.

This adds external storage permission and will prompt users for read access for photos / videos etc. when they click on something that uses the imagePicker. Even though we cannot reliably reproduce this, it should fix the problem.

Fixes #5792

* Address feedback