Load assignments synchronously. (#5840)
Load assignments synchronously at startup time. 

Measurements have shown that this implementation is fast enough to be done on the main thread at startup time. This removes all race conditions and ensures the experiment configuration is loaded asap. It has also the advantage of being more consistent with PlanStorage and ConfigurationSettingsStorage.