Log background syncing in advance of migrating (#6000)
We are going to migrating to new background syncing logic soon. It would be helpful to be able to tell quickly if the new logic was performing in line with the old logic. In order to do that, lets send an event to fabric overtime we do a background sync with some data about how much is being synced

Also fixed a bug where we would not set the priority field from the background, which would have resulted in things defaulting to high priority in the new synced.