Add warnings for scheduler bottlenecks (#6130)
* Add warnings for scheduler bottlenecks

This is to help us map out problems that we are seeing with betas. This way we will see warnings when there are excessive wait times in schedulers, or if the queue size gets too large.

* Add other option feedback item in menu

Since we are now requesting logs from users for non playback related things, this gives them the ability to send one explicitly, so the playback team can ignore the email.