Fix null cache path handling in file access event (#6209)
Refs #6203

It is possible for the cache directory to not exist (if the disk is mounted for example). In this case, make sure we handle null.

Note, this will just send a [does not exist | is not readable | is not writeable] event in this case. Not sure if that is ok, but just making this not crash for the release.