Refactor TrackLikesHeaderPresenter to be clearer (#6176)
I need to do some work on this presenter for adding a filter to the top. I realised it was a bit unclear how the data flowed, and when this flow resulted in a change. I tried to make this clearer by using ViewModels and subjects. I actually think this is a good example of how we should do presenters (that are not collections).

Feedback appreciated!