Do not sync offline playlists from collection foreground (#6117)
I notice that collection syncing takes a long time if you have a lot of offline playlists. This is because we sync all of them everytime you request a sync.

To fix this, do not sync every offline playlist when requesting a sync from the foreground. Only do it from a background sync.

To achieve this, I had to post the request origin through to the sync provider.

I also bumped autoFactory, so we could create factories with named in section. This forced me to shuffle a couple classes around, but should result in no major functional changes.