Only notify the item that changed in our subscribers (#6285)
By taking advantage of the fact that we use RecyclerViews, we can minimise the updates necessary when responding to our eventBus.

There is one shortcoming here, in that by default, notifyItemChanged/Added/Removed will cause an animation on the lists. Because we do not have a need for custom animations here, I chose to disable the animations on the recycler views to avoid the “blinking” that occurs when the same image animates into itself. If we want a different strategy we can discuss, or we can pull this into android kit (no animations by default) so we don’t have to remember to do it.