fix flaky acceptancetest PlayerTest_testPlayerShowTheTrackNoDescription (#6432)
* fix flaky acceptancetest PlayerTest_testPlayerShowTheTrackNoDescription

The test was flaky since it was searching for the content of the view
R.id.title. This id is valid for many views visible during that testrun,
including the PopupMenu items, which all have their TextView with the id:
R.id.title. It seems like a race condition that sometimes the Track Info Dialog
doesn't show up fast enough and the Test Runner picks up the first PopupMenu
item as the title it checks against.

This could probably also have been solved with a small sleep, however I think
that giving the view a unique name is much clearer.

* remove outdated todo in tests