Put track likes header inside list (#6304)
This is prep for adding a filter above the header.

You will notice there is an intermediate view model / data source class in the track likes presenter now. This is the cleanest way I have found to provide both paging, and header items inside lists.

The main reason is that headers should only be added to the first page of a list. In order to achieve this, I tried a couple of things

- Stateful transformer, that would only add the header to the first emission. This does not work, as the observables are replayed, so you may emit the first page multiple times. The Transformer does not know that, so this will not work.

- Prepending an extra item (the header) to the initial observable sequence. This worked, but ultimately was a bit weird when it came to paging, as the pager had to basically throw out the header item when deciding if it should keep paging or not.

The solution I settled on was the new data object which encodes the header presence, without making the pager jump through hoops. Because the pager is now coupled to the data type in the presenter, I decided to pull all of this logic into a DataSource object (as we do in Discovery), for organisational purposes