Acceptance Tests: merge TrackLikesTests to one single test (#6449)
* merge TrackLikesTests to one single test

by merging these tests we speed up the test suite as it doesn't have to
go to the likes page for every single of the tests before executing any
assertions.

Second: `testPlayAndPauseTrackFromLikes` was flaky because it started playing
a track and tried to pause it and we verify the tracking. However the very first
track in this users likes was only 8s long. The play and pause actions sometimes
took longer than 8s so a second track was started and tracking results were invalid
since it contained playback events from the second test. By now scrolling down
the list before starting to play we're hitting a longer track that should make this
test less flaky as long as the test data doesn't change.

* verify that liked track is >1min before playing

* inline variable in TrackLikesScreen