[PLAYBACK-1673] Fix ghost player behavior (#6535)
We were calling showPlayer when connecting to the cast even with a empty remote play queue. The correct implementation will only show it if there is actual content.

Most of the changes on this PR are related to renaming stuff; making RemotePlayQueue an @Autovalue model class (and encapsulating some logic there) and introducing Lambdas. You should look into DefaultCastPlayer for the actual bugfix.