Remaining Profile Design Tweaks (#6541)
* Remaining Profile Design Tweaks

- fix back arrow positioning
- fix icon colouring based on scroll position
- some cleanup

Note: Because of status bar icon coloring, we only enable this on M and above. Sorry late adopters!

* Added StatusBarColorController to fix color conflicts

Because we are colouring the status bar in more than one place, we have to have a central place where we can manage this so that the proper state is restored. This new class comes in to properly manager the colouring with regards to player status, and to restore then proper states when the player is closed.