Stream clear adapter on new items (#6809)
Temporary fix so we insert ads on the main thread. Refs #6807

Clear ads on initial load of page instead of before

When you would press the bubble to show new items, we were previously calling clear adapter without notifying the adapter. At some point recently, this became an issue, possibly as a side effect of something we were doing with ads.

Fix this by always clearing the adapter when we load initial items. This should result in no change in functionality.