AS-1035 Use Firebase Dynamic Links when sharing.
http://go/j/AS-1035

Today, when a user shares a track/playlist/whatever from the app,
we create a share message like this:

  Listen to Touchdown by BETSHEWILLZ #np on #SoundCloud
  https://soundcloud.com/betshewillz/touchdown

This change introduces a new feature flag that, when enabled, will
cause us to create share messages like this:

  Listen to Touchdown by BETSHEWILLZ #np on #SoundCloud
  https://soundcloud.app.goo.gl/BCF5?/betshewillz/touchdown

The URL is now a Firebase Dynamic Link. The advantage is that users
without the app will be taken to the Play or App store. Once they have
installed the app and signed in, they will be taken to the original
content.

See https://firebase.google.com/docs/dynamic-links/

Note that if the request to Firebase fails or times out, we
fall back to the original URL.

Caveats:
- We still need design guidance on how we handle the delay we are
  introducing between clicking share and seeing the share dialog.
- We still need to update the link URLs from g2waz.app.goo.gl to
  soundcloud.app.goo.gl. Naom is working on this.