Allow up to 1000 compiler errors instead of 100 (#6937)
When we have a mistake in our code generation, it will result in a compiler error at every point we depend on generated code. This can make it impossible to find the real error, as the compiler will only report a maximum of 100 errors. By increasing this to 1000, we have enough head room to see all errors.

Note: 1000 is a bit of a magic number. We currently use AutoValue at roughly 150 places. 1000 errors will allow us a 6x usage increase of auto value while still reporting all errors, so we should be safe. We could potentially set this to a number that we will never reach, but I am unsure of potential side effects of doing so, so I would like to do this incrementally.