Collection Loader for re-usable MVI collections (#7029)
In the quest to be more architecturally compliant, while still writing re-usable code, I realised we needed to come up with something that can handle generic loading / paging / refreshing in a way that was consistent and re-usable for most of our collections. This is my attempt at that.

The idea is that it will respond to emissions from users (load/refresh/page) output a typed collection model, that we can then transform to a view model, and can be understood in a re-usable UI component (list fragment).

Though this is not used yet, I would like to merge it in being unused if there are no major objections, so that I may introduce it in a separate PR on existing work.