Don't reuse OkHttpClient.Builders. (#7168)
We were providing a singleton OkHttpClient.Builder in ApiModule, then
injecting in various other modules to allow clients to be created with
custom settings. The problem is that if any code modifies that shared
builder, all subsequent uses will be impacted. The documented way of
sharing a connection pool across clients is to use
OkHttpClient.newBuilder.