Remove AppBarLayout from playlist details (#7161)
Because google can’t get the *&@# together and fix the support library, we have to go back to using a normal recycler view, and manually handle all of the header expansions and animations. We do this in a new collaborator,

This naturally involves moving all of our header rendering logic to the adapter. Luckily this was not too difficult because of previous refactors.

There were also so small refactors and polishes along with this, such as

- extending the hit area of the drag handles
- making the status bar gray instead of white when opaque
- made scroll speed of recycler view a bit slower as it was too jarring

Also note : there are very few tests with this change, as it is mostly view and animation logic, and very volatile. I don’t think it is worth testing.