Introduce `assertIsEnabled` in `FeatureFlagsHelper`. (#7211)
When the feature flag is used when booting the app (`bootApplication`) the test cannot enabled the feature flag (the read already happened when the test runner started the app) and the test has to rely on the grade configuration. 

This assertion will be used to guarantee the required flags are enabled before starting the test.