Ignore IDLE state handling after load command has been issues for a session (#7260)
As part of the custom protocol, we are establishing that a given connected session will
NEVER broadcast an IDLE state after the first load command has been issued by a sender.
The SDK, though, might internally send some of those states if the connection gets
flaky or into a transient state. We shall ignore those in order to provide a more
“fluid” experience to our users since we don’t want the UI to transition between states
as much as the underlying protocol does.