Introduce PlaybackProgress in-memory cache (#7348)
* Introduce progress in-memory cache based on URN

Every progress updated must be redirected to this class, as it will hold the
shared state of progress for all played tracks.

* Some minor refactoring

* Publish empty progress when clearing it from the storage

* Update variable name reflecting its class

* Do not forcefully update progress onTrackChange

Will only do that via the Queue subscriber. Let’s not have different places updating the
progress state as that introduces a race condition and might lead to out-of-sync data
on the UI.

* Avoid going to the database on every position update

All we need from the database is the duration of the track. If that URN is already
on the cache, there’s no need to lookup the database another time, let’s just merge
that value with the updated position and set that on the cache.

* Use ConcurrentHashMap to avoid threading issues while read/writing from the storage

* Replace if-else with fancy Optional magic